升级依赖包

+++ web/models/admin_import_menu.go
@@ -11,7 +11,7 @@ type AdminMenuImport struct {
 	MenuModule    string     `json:"menu_module" gorm:"column:menu_module"`
 	ImportAppName string     `json:"import_app_name" gorm:"column:import_app_name"`
 	ImportId      int64      `json:"import_id" gorm:"column:import_id"`
-	DefaultOpen   int        `json:"default_open" gorm:"column:default_open"` // 默认开启菜单就开启本接口权限 1-开启 不为1-不开启（ 默认值2不开启）
+	DefaultOpen   uint8      `json:"default_open" gorm:"column:default_open"` // 默认开启菜单就开启本接口权限 1-开启 不为1-不开启（ 默认值2不开启）
 	CreatedAt     time.Time  `json:"-" gorm:"column:created_at"`
 	UpdatedAt     time.Time  `json:"-" gorm:"column:updated_at"`
 	DeletedAt     *time.Time `json:"-" gorm:"column:deleted_at"`
